requirements.txt
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pyhd8ed1ab_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=py_2
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
black=22.1.0=pypi_0
blas=1.0=mkl
bleach=4.1.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.10.26=h06a4308_2
certifi=2021.10.8=py39h06a4308_2
cffi=1.15.0=pypi_0
click=8.0.4=pypi_0
colorama=0.4.4=pypi_0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pypi_0
cython=0.29.28=pypi_0
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pypi_0
docopt=0.6.2=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
executing=0.8.2=pyhd8ed1ab_0
ffmpeg=4.3=hf484d3e_0
fonttools=4.29.1=pypi_0
freetype=2.11.0=h70c0345_0
genco=1.0.0=dev_0
genco-tsp=1.0.0=dev_0
giflib=5.2.1=h7b6447c_0
gitdb=4.0.9=pypi_0
gitpython=3.1.27=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
iniconfig=1.1.1=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.9.1=py39hef51801_0
ipython=8.0.1=py39hf3d152e_1
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.6.5=pypi_0
jedi=0.18.1=py39hf3d152e_0
jinja2=3.0.3=pypi_0
joblib=1.1.0=pypi_0
jpeg=9d=h7f8727e_0
jsonpickle=1.5.2=pypi_0
jsonschema=4.4.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-console=6.4.0=pypi_0
jupyter_client=7.1.2=pyhd8ed1ab_0
jupyter_core=4.9.2=py39hf3d152e_0
jupyterlab-pygments=0.1.2=pypi_0
jupyterlab-widgets=1.0.2=pypi_0
kiwisolver=1.3.2=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h295c915_1
markupsafe=2.1.0=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pyhd8ed1ab_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
munch=2.5.0=pypi_0
mypy-extensions=0.4.3=pypi_0
nbclient=0.5.11=pypi_0
nbconvert=6.4.2=pypi_0
nbformat=5.1.3=pypi_0
ncurses=6.3=h7f8727e_2
nest-asyncio=1.5.4=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pyhd3eb1b0_0
notebook=6.4.8=pypi_0
numpy=1.22.2=pypi_0
olefile=0.46=pyhd3eb1b0_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1m=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.4.1=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pathspec=0.9.0=pypi_0
pexpect=4.8.0=pyh9f0ad1d_2
pickleshare=0.7.5=py_1003
pillow=8.4.0=py39h5aabda8_0
pip=21.2.4=py39h06a4308_0
platformdirs=2.5.1=pypi_0
pluggy=1.0.0=pypi_0
prometheus-client=0.13.1=pypi_0
prompt-toolkit=3.0.27=pyha770c72_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
py=1.11.0=pypi_0
py-cpuinfo=8.0.0=pypi_0
pyconcorde=0.1.0=dev_0
pycparser=2.21=pypi_0
pygments=2.11.2=pyhd8ed1ab_0
pymongo=3.9.0=pypi_0
pyparsing=3.0.7=pypi_0
pyrsistent=0.18.1=pypi_0
pytest=7.0.1=pypi_0
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pyhd8ed1ab_0
python-sat=0.1.7.dev15=pypi_0
python_abi=3.9=2_cp39
pytorch=1.10.2=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2021.3=pypi_0
pyyaml=6.0=pypi_0
pyzmq=19.0.2=py39hb69f2a1_2
qtconsole=5.2.2=pypi_0
qtpy=2.0.1=pypi_0
readline=8.1.2=h7f8727e_1
sacred=0.8.2=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.8.0=pypi_0
seml=0.3.6=pypi_0
send2trash=1.8.0=pypi_0
setuptools=58.0.4=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
smmap=5.0.0=pypi_0
sqlite=3.37.2=hc218d9a_0
stack_data=0.2.0=pyhd8ed1ab_0
terminado=0.13.1=pypi_0
testpath=0.5.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.11=h1ccaba5_0
tomli=2.0.1=pypi_0
torch-scatter=2.0.9=pypi_0
torch-sparse=0.6.12=pypi_0
torchaudio=0.10.2=py39_cu113
torchtyping=0.1.4=pypi_0
torchvision=0.11.3=py39_cu113
tornado=6.1=py39h3811e60_1
tqdm=4.62.3=pypi_0
traitlets=5.1.1=pyhd8ed1ab_0
typeguard=2.13.3=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
wcwidth=0.2.5=pyh9f0ad1d_2
webencodings=0.5.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.5.2=pypi_0
wrapt=1.13.3=pypi_0
xz=5.2.5=h7b6447c_0
zeromq=4.3.4=h9c3ff4c_0
zlib=1.2.11=h7f8727e_4
zstd=1.4.9=haebb681_0
